WeeklyCalendar: move NextArrow & PrevArrow inside this component
