+++ src/containers/EditListingPage/EditListingWizard/EditListingAvailabilityPanel/WeeklyCalendar/NextArrow.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { isDateSameOrAfter } from '../../../../util/dates';
-import { IconArrowHead } from '../../../../components';
+import { isDateSameOrAfter } from '../../../../../util/dates';
+import { IconArrowHead } from '../../../../../components';
 
 // Component for the react-dates calendar
 const NextArrow = props => {
